+++ CHANGELOG.md
@@ -1,1 +1,11 @@
 # Lintje changelog
+
+## 0.1.1
+
+- Add MessageEmptyFirstLine rule that checks if the line after the subject line
+  is empty. If it's not empty that line is considered part of the commit's
+  subject.
+
+## 0.1.0
+
+Initial release.